adjusted navbar, footer and other page components to make app mobile friendly
