+++ client/components/Admin/Users/AdminUsersTable.js
@@ -93,7 +93,7 @@ const UsersTable = (props) => {
   /////////////////////
 
   return (
-    <Container sx={{ mt: 3, marginBottom: 25 }}>
+    <Container sx={{ mt: 3, pb: 25 }}>
       <Button variant="outlined" component={Link} to={`/admin/users/create`}>
         Create New User
       </Button>
