--- client/components/Materialui/Buttons/Icon-Button.js
@@ -1,13 +1,0 @@
-import { IconButton } from '@mui/material'
-import React from 'react'
-
-const StandardButton = (props) => {
-    const { color, icon } = props
-    return (
-        <IconButton color={color || "primary"}>
-            {icon}
-        </IconButton>
-    )
-}
-
-export default StandardButton
