+++ client/components/Products/ProductCard.js
@@ -46,25 +46,25 @@ function ProductCard(props) {
   };
 
   return (
-      <Box component={Link} to={`/products/${id}`} >
-        <Card raised sx={{ maxWidth: 350, boxShadow: 'none', '&:hover': { boxShadow: '0px 5px 5px -3px rgb(0 0 0 / 20%), 0px 8px 10px 1px rgb(0 0 0 / 14%), 0px 3px 14px 2px rgb(0 0 0 / 12%)' } }}>
-          <CardMedia component="img" height="400" image={imageURL} sx={{ p: 1 }} />
-          <CardContent sx={{ textAlign: "center" }}>
-            <Typography variant="body2" color="text.secondary">
-              {brandName}
-            </Typography>
-            <Typography variant="h6">
-              {name}
-            </Typography>
-            <Rating name="half-rating-read" defaultValue={rating * 1} precision={0.5} readOnly />
-            <Typography variant="body2" color="text.secondary">
-              Price: {currency(price).format()}
-            </Typography>
-          </CardContent>
-          <CardActions sx={{ display: "flex", flexDirection: "column" }}>
-          </CardActions>
-        </Card>
-      </Box>
+    <Box component={Link} to={`/products/${id}`} >
+      <Card raised sx={{ maxWidth: 350, boxShadow: 'none', '&:hover': { boxShadow: '0px 5px 5px -3px rgb(0 0 0 / 20%), 0px 8px 10px 1px rgb(0 0 0 / 14%), 0px 3px 14px 2px rgb(0 0 0 / 12%)' } }}>
+        <CardMedia component="img" height="400" image={imageURL} sx={{ p: 0 }} />
+        <CardContent sx={{ textAlign: "center", p: 0 }}>
+          <Typography variant="body2" color="text.secondary">
+            {brandName}
+          </Typography>
+          <Typography variant="h6">
+            {name}
+          </Typography>
+          <Rating name="half-rating-read" defaultValue={rating * 1} precision={0.5} readOnly />
+          <Typography variant="body2" color="text.secondary">
+            Price: {currency(price).format()}
+          </Typography>
+        </CardContent>
+        <CardActions sx={{ display: "flex", flexDirection: "column" }}>
+        </CardActions>
+      </Card>
+    </Box>
   );
 }
