+++ public/style.css
@@ -1,7 +1,8 @@
 body {
-  font-family: sans-serif;
+  position: relative;
+  min-height: 100vh;
 }
-
+/* 
 a {
   text-decoration: none;
 }
@@ -13,5 +14,4 @@ label {
 nav a {
   display: inline-block;
   margin: 1em;
-}
-
+} */
